Use ReactQuery and remove state
