--- client/state/promote-post/init.ts
@@ -1,4 +1,0 @@
-import { registerReducer } from 'calypso/state/redux-store';
-import reducer from './reducer';
-
-registerReducer( [ 'promotePost' ], reducer );
